requirements.txt
Django==4.2.4
djangorestframework==3.14.0
django-cors-headers==4.2.0
django-crispy-forms==1.14.0
crispy-bootstrap5==0.7
whitenoise==6.2.0
psycopg2==2.9.7
psycopg2-binary==2.9.7
redis==4.5.4
bcrypt==4.0.1
cryptography==41.0.5
argon2-cffi==23.1.0
gunicorn==20.1.0
python-dotenv==1.0.0
djangorestframework-simplejwt==5.3.0
PyJWT==1.7.1
requests==2.31.0
pandas==2.2.2
numpy==1.23.5
pytz==2023.3
six==1.16.0